SudokuPromptFlowDocker/flow/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
azure-core==1.30.1
azure-identity==1.15.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Deprecated==1.2.14
diffusers==0.26.3
distro==1.9.0
docutils==0.20.1
filelock==3.13.1
filetype==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
flask-restx==1.3.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-auth==2.28.2
google-search-results==2.4.1
googleapis-common-protos==1.63.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.3.1
itsdangerous==2.1.2
jaraco.classes==3.3.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==24.3.1
MarkupSafe==2.1.3
marshmallow==3.21.1
more-itertools==10.2.0
mpmath==1.2.1
msal==1.27.0
msal-extensions==1.1.0
networkx==3.2.1
numpy==1.26.4
openai==1.14.1
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
pandas==2.2.1
pillow==10.2.0
portalocker==2.8.2
prettytable==3.10.0
promptflow==1.6.0
promptflow-tools==1.3.0
protobuf==4.25.3
pyaes==1.6.1
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydash==7.0.7
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.28
strictyaml==1.7.3
sympy==1.12
tabulate==0.9.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.66.2
transformers==4.38.2
tzdata==2024.1
urllib3==2.2.1
waitress==2.1.2
Werkzeug==3.0.1
wrapt==1.16.0